+++ packages/apps/kadena-docs/next.config.js
@@ -5,8 +5,12 @@ import remarkHeadersToProps from './src/scripts/remarkHeadersToProps.js';
 import remarkSideMenuToProps from './src/scripts/remarkSideMenuToProps.js';
 import remarkPropsToStaticRender from './src/scripts/remarkPropsToStaticRender.js';
 import remarkAdmonitions from './src/scripts/remarkAdmonitions.js';
+import remarkYoutube from './src/scripts/remarkYoutube.js';
+import remarkFigureOutOfParagraph from './src/scripts/remarkFigureOutOfParagraph.js';
 import remarkGfm from 'remark-gfm';
 import mdx from '@next/mdx';
+import { createVanillaExtractPlugin } from '@vanilla-extract/next-plugin';
+const withVanillaExtract = createVanillaExtractPlugin();
 import { getHighlighter, BUNDLED_LANGUAGES } from 'shiki';
 import { readFileSync } from 'fs';
 
@@ -70,6 +74,8 @@ const withMDX = mdx({
       remarkSideMenuToProps,
       remarkPropsToStaticRender,
       remarkAdmonitions,
+      remarkYoutube,
+      remarkFigureOutOfParagraph,
     ],
     rehypePlugins: [[rehypePrettyCode, options]],
     // If you use `MDXProvider`, uncomment the following line.
@@ -81,6 +87,7 @@ const withMDX = mdx({
 const nextConfig = {
   pageExtensions: ['ts', 'tsx', 'js', 'jsx', 'md', 'mdx'],
   reactStrictMode: true,
+  transpilePackages: ['@kadena/react-ui'],
 };
 
-export default withMDX(nextConfig);
+export default withVanillaExtract(withMDX(nextConfig));