+++ packages/apps/kadena-docs/src/components/Markdown/MDNotification/MDNotification.tsx
@@ -1,4 +1,5 @@
 import { Notification, SystemIcons } from '@kadena/react-components';
+import { NotificationBody } from '@kadena/react-components/types/components/Notification';
 
 import React, { FC, ReactNode } from 'react';
 
@@ -7,9 +8,13 @@ interface IProps {
 }
 
 export const MDNotification: FC<IProps> = ({ children, ...props }) => {
+  console.log({ children, props });
   return (
     <div {...props}>
-      <Notification title="test" expand icon={SystemIcons.Account} />
+      1{children}1
+      <Notification title="test" expand icon={SystemIcons.Account}>
+        <NotificationBody>{children}</NotificationBody>
+      </Notification>
     </div>
   );
 };