+++ packages/apps/kadena-docs/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './Breadcrumbs';
 export * from './BrowseSection';
 export * from './LinkList';
 export * from './Analytics';
+export * from './LastModifiedDate';