+++ packages/libs/react-components/.eslintrc.js
@@ -4,9 +4,4 @@ require('@rushstack/eslint-config/patch/modern-module-resolution');
 module.exports = {
   extends: ['@kadena-dev/eslint-config/profile/react'],
   parserOptions: { tsconfigRootDir: __dirname },
-  settings: {
-    react: {
-      version: 'detect',
-    },
-  },
 };