+++ packages/tools/eslint-config/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @kadena-dev/eslint-config
 
-This log was last generated on Fri, 03 Mar 2023 11:24:59 GMT and should not be manually modified.
+This log was last generated on Thu, 01 Jun 2023 20:18:44 GMT and should not be manually modified.
+
+## 0.0.5
+Thu, 01 Jun 2023 20:18:44 GMT
+
+### Patches
+
+- Updated React profile to detect react version
 
 ## 0.0.4
 Fri, 03 Mar 2023 11:24:59 GMT